Auth cleanup
This makes several cleanups in the implementation of roles and auth.

Currently, `cfg::Config.auth` entries are modeled on PostgreSQL's HBA
table, and include the database field.  In my opinion, the "database"
spec is misplaced.  It makes little sense to customize user
authentication methods on a per-database basis, and the actual *ability*
to connect to a certain database is an *authorization* issue, not an
*authentication* one.  This also removes the `host` attribute, since we
don't do host-based authentication yet.

Additionally, the `CREATE ROLE` command currently creates accounts that
are, effectively, superusers, since we currently lack the facilities to
grant granular access permissions.  To avoid future backwards
incompatibility the current `CREATE ROLE` command and the corresponding
CLI must ensure that the superuser intent is explicitly specified.
Although the current `CREATE ROLE` command has the
`SET is_superuser := <true|false>` operation, it is a) unwieldy to type;
b) does not emphasize the special nature of the new role.  Hence, this
adds the `CREATE SUPERUSER ROLE` syntax for this purpose.
We already have precedent in elevating important object qualifiers to
the prefix keyword status (such as `CREATE REQUIRED LINK`).  The
corresponding CLI command is `create-superuser-role`.

Finally, the `allow_login` property of the `Role` is now deemed dubious
and is removed.  First of all, the majority of roles created are
normally allowed to log in, so making it `True` by default probably
makes more sense, and then the presence of a separate login flag seems
superfluous with respect to the Auth records.  The only valid reason why
the flag might be used is to (temporarily) disable the login capability,
and a `Deny` auth method or a similar mechanism seems more suited for
that.  Using "allow_login" to distinguish between roles that are
"groups" and roles that are "users" seems misplaced, and if we ever do
need that, a qualifier like `ABSTRACT` would seem to be more appropriate.

Fixes: #1128.
Fixes: #1129.