Make sys::get_version() return the actual version of the running server
Currentlly, `sys::get_version()` returns the version of of a server that
was used to *bootstrap* an instance, since version metadata is written
into the instancedata storage.  This is obviously wrong, since one might
run a different, but compatible, version of EdgeDB on the same data
instance.  Fix this by switching `sys::get_version()` to use the session
state table instead.